server/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bike-recorder-server"
version = "0.1.0"
description = "BikeRecorder backend API"
authors = [{name = "BikeRecorder", email = "dev@bikerecorder.local"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110,<0.112",
    "uvicorn[standard]>=0.29,<0.30",
    "sqlmodel>=0.0.16",
    "pydantic-settings>=2.2,<2.3",
    "pyjwt>=2.8,<3.0",
    "python-multipart>=0.0.9,<0.0.10",
    "httpx>=0.27,<0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1,<8.2",
    "anyio>=4.2,<4.4",
]
